We are now unblocked, ready to move forward with some positive energies
